requirements.txt
flask
docx-mailmerge
fitz
PyMuPDF
pypandoc
pyodbc
